+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [next]
 - Bugfix/quadtree id for removal [#178](https://github.com/RafaelBarbosatec/bonfire/pull/178)
+- Adds `keyboardDirectionalType` param in `Joystick` to enable WASD.
 
 # 1.7.0
 - adds `FollowerWidget`. With this you can add a widget what follows a component in the game.